Rename write config phase
